Removed superfluous field conversion
